+++ src/main/java/insane96mcp/survivalreimagined/module/mining/client/MultiBlockSoulBlastFurnaceScreen.java
@@ -1,0 +1,14 @@
+package insane96mcp.survivalreimagined.module.mining.client;
+
+import insane96mcp.survivalreimagined.SurvivalReimagined;
+import insane96mcp.survivalreimagined.module.mining.inventory.MultiBlockSoulBlastFurnaceMenu;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
+
+public class MultiBlockSoulBlastFurnaceScreen extends AbstractMultiBlockFurnaceScreen<MultiBlockSoulBlastFurnaceMenu>{
+    private static final ResourceLocation TEXTURE = new ResourceLocation(SurvivalReimagined.MOD_ID, "textures/gui/container/blast_furnace.png");
+    public MultiBlockSoulBlastFurnaceScreen(MultiBlockSoulBlastFurnaceMenu pMenu, Inventory pPlayerInventory, Component pTitle) {
+        super(pMenu, new MultiBlockSoulBlastFurnaceRecipeBookComponent(), pPlayerInventory, pTitle, TEXTURE);
+    }
+}